+++ frontend/src/routes/login/+page.svelte
@@ -1,0 +1,1 @@
+Login to Quibble
